Made better use of existing code
