+++ vispy/gloo/context.py
@@ -19,9 +19,9 @@ an OpenGL context.
 from copy import deepcopy
 import weakref
 
-from .. import glir_file
 from .glir import GlirQueue, BaseGlirParser, GlirParser, glir_logger
 from .wrappers import BaseGlooFunctions
+from .. import config
 
 _default_dict = dict(red_size=8, green_size=8, blue_size=8, alpha_size=8,
                      depth_size=16, stencil_size=0, double_buffer=True,
@@ -161,7 +161,8 @@ class GLShared(object):
     # left, things should Just Work. 
     
     def __init__(self):
-        parser_cls = GlirParser if (glir_file is None) else glir_logger(GlirParser, glir_file)
+        glir_file = config['glir_file']
+        parser_cls = glir_logger(GlirParser, glir_file) if glir_file else GlirParser
         self._parser = parser_cls()
         self._name = None
         self._refs = []
